Added user/admin actions logging by AOP advices

+++ src/main/java/io/leonid/springcash/model/BaseEntity.java
@@ -10,14 +10,14 @@ public abstract class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    protected Integer id;
+    protected Long id;
 
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 }
